feat: add validation schemas for inquiries and packages using Zod refactor: enhance API functions with error handling

+++ app/api/inquiries/[id]/route.ts
@@ -2,7 +2,9 @@ import { NextResponse } from "next/server";
 import connectMongoDB from "@/lib/mongodb";
 import Inquiry from "@/modals/Inquiry";
 import Package from "@/modals/Package";
+import { inquirySchema } from "@/lib/types";
 
+// Get a inquiry by id
 export async function GET(
 	request: Request,
 	{ params }: { params: Promise<{ id: string }> }
@@ -10,25 +12,54 @@ export async function GET(
 	const { id } = await params;
 	await connectMongoDB();
 	Package.init();
-	const data = await Inquiry.findById(id).populate("packageId", "name").lean();
-	return NextResponse.json(data, { status: 200 });
+
+	const inquiry = await Inquiry.findById(id).populate("packageId", "name");
+
+	if (!inquiry) {
+		return new NextResponse(null, { status: 404 });
+	}
+
+	return NextResponse.json(inquiry, { status: 200 });
 }
 
+// Update a inquiry by id
 export async function PUT(
 	request: Request,
 	{ params }: { params: Promise<{ id: string }> }
 ) {
 	const { id } = await params;
-	const updatedInquiry = await request.json();
-	await Inquiry.findByIdAndUpdate(id, updatedInquiry);
-	return NextResponse.json(updatedInquiry, { status: 201 });
+	const updatedInquiry = inquirySchema.safeParse(await request.json());
+
+	if (!updatedInquiry.success) {
+		return NextResponse.json(updatedInquiry.error.issues, { status: 400 });
+	}
+
+	await connectMongoDB();
+
+	const inquiry = await Inquiry.findById(id);
+
+	if (!inquiry) {
+		return new NextResponse(null, { status: 404 });
+	}
+
+	await Inquiry.findByIdAndUpdate(id, updatedInquiry.data);
+	return NextResponse.json(updatedInquiry.data, { status: 201 });
 }
 
+// Delete a inquiry by id
 export async function DELETE(
 	request: Request,
 	{ params }: { params: Promise<{ id: string }> }
 ) {
 	const { id } = await params;
+	await connectMongoDB();
+
+	const inquiry = await Inquiry.findById(id);
+
+	if (!inquiry) {
+		return new NextResponse(null, { status: 404 });
+	}
+
 	await Inquiry.findByIdAndDelete(id);
 	return new NextResponse(null, { status: 204 });
 }
